latest

+++ notes_server/routes/notes.js
@@ -3,7 +3,6 @@ const Note = require("../models/Notes");
 const User = require("../models/User");
 
 //create notes
-
 router.post("/addNotes", async (req, res) => {
   try {
     const note = new Note({
@@ -24,26 +23,44 @@ router.post("/addNotes", async (req, res) => {
 router.delete("/deleteNotes/:id", async (req, res) => {
   try {
     const notes = await Note.findOne({ _id: req.params.id });
-    !notes &&
-      res.status(400).json({ message: "note not found ", status: false });
-    const note = await Note.deleteOne({ _id: req.params.id });
-    res.status(200).json({ message: "note deleted successfull", status: true });
+    if (!notes) {
+      return res.status(400).json({ message: "Note not found", status: false });
+    }
+
+    const result = await Note.deleteOne({ _id: req.params.id });
+
+    if (result.deletedCount > 0) {
+      res
+        .status(200)
+        .json({ message: "Note deleted successfully", status: true });
+    } else {
+      res
+        .status(500)
+        .json({ message: "Error deleting the note", status: false });
+    }
   } catch (e) {
     res.status(500).json(e);
   }
 });
-
 //update notes
 router.put("/updateNotes/:id", async (req, res) => {
   try {
     const notes = await Note.findOne({ _id: req.params.id });
-    !notes &&
-      res.status(400).json({ message: "note not found ", status: false });
-    const note = await Note.updateOne({
-      title: req.body.title,
-      description: req.body.description,
-      postedBy: req.body.postedBy,
-    });
+    if (!notes) {
+      return res.status(400).json({ message: "note not found", status: false });
+    }
+
+    const result = await Note.updateOne(
+      {
+        _id: req.params.id,
+      },
+      {
+        title: req.body.title,
+        description: req.body.description,
+        postedBy: req.body.postedBy,
+      }
+    );
+
     res
       .status(200)
       .json({ message: "note updated successfully", status: true });
@@ -57,7 +74,7 @@ router.get("/getNotes/:userId", async (req, res) => {
   try {
     const currentUser = await User.findById(req.params.userId);
 
-    !currentUser && res.status(400).json("user not found");
+    !currentUser && res.status(400).json({ message: "user not found" });
     const notes = await Note.find({ postedBy: req.params.userId });
 
     res.status(200).json(notes);
